Set drives to unsafe mode
Default to qemu "unsafe" mode for image files which turns of cache
flushing into the image file. This makes the i/o load much lower on the
host if e.g. install debian packages which triggers fdatasync() all the
time.

In almost all use-cases if the host crashes the image files would be
rebuild anyway rather then assumed to be in a good state, so doing this
by default seems reasonable.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>