Use an explicit fakemachine.service
The fakemachine runner is launched by overriding the
serial-getty@ttyS0.service file, relying on the systemd getty generator
to add that a Wants relationship to default.target for it.

Make this all a bit less magic by instead generating
fakemachine.service and adding it as the target systemd unit on the
kernel command line. This has the added benefit that fakemachine will
only start the units needed directly rather then triggering arbitrary
services that might be part of default.target

Just to be sure there are no tty conflicts, mask serial-getty@ttyS0 to
prevent that from being started in all cases.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>