Let the command wrapper handling network waiting
We've seen some cases where for reasons that aren't yet clear networkd
fails to start on fakemachine boot. Currently this means that the
dependencies for the fakemachine runner fail, which causes it to never
be started and fakemachine to just hang.

Make this a bit more robust by having the command wrapper script be the
one to wait on the network to come online and, in case of failure, dump
out diagnostics. In this case systemd-networkd because a Wants rather
then a Requires such that if that fails to start the fakemachine service
still gets run.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>