Disable terminal signal handling
As a side-effect of changing how boot messages get handled the signal
handling on the terminal got turned on (qemu default). This means that
when running a command in fakemachine Control-c went to qemu not the
command in fakemachine. This is rather annoying when using e.g. bash in
fakemachine.

This patch restores the previous behaviour, although this probably
should be configurable at some point.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>