machine: Refine the logic to find the kernel release
Currently, the kernel release is determined by taking the last entry in
/usr/lib/modules. This doesn't work for distributions like ArchLinux,
which has additional entries in /usr/lib/modules:

    $ ls -1 /usr/lib/modules
    5.2.4-arch1-1-ARCH
    extramodules-ARCH

This commit fixes it. We keep the same logic (ie. pick up the last
kernel found in /usr/lib/modules), however we add an additional
constraint: filter out things that don't start with a digit (because
yes, we strongly expect that the kernel release starts with a digit).

(this neat trick was taken from the mkosi source code)

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>