Support virtfs and mount options per volume
For some use cases, users may wish to use the mapped-file or mapped-xattr
security_models for some volumes; similarly, they may wish to disable 9p
caching.  This change extends the colon-delimited expression accepted by
the "-v" flag to include an optional third component: a comma-delimited
list of virtfs and mount options, each prefixed with "virtfs." or
"mount.", e.g.:

    fakemachine \
      -v path:/path:virtfs.security_model=mapped-file,mount.cache=none

    results in this argument to qemu:
        -virtfs ...,security_model=mapped-file

    and this line in the fakemachine's fstab:
        virtfs-1 ... ...cache=none,msize=262144 0 0

This is backwards-compatible in both the CLI and library interfaces.  If
the aforementioned third component is omitted, current behavior is maintained.
On the library side, a variadic argument is added to (*Machine).AddVolumeAt;
the behavior of existing callers is unchanged.