machine: Rework the logic to find libraries (libc and libresolv)
We used to look for these libraries in (usr)/lib/x86_64-linux-gnu/, but
this is location is Debian-specific, other distros use different paths.

For example, ArchLinux puts everything in /usr/lib, while Fedora uses
/lib64 apparently.

With this commit, we assume that the libraries are installed in the same
directory as the linker (which itself is installed at a well known-path
per architecture as defined by [1]). This is true for ArchLinux, Fedora,
Debian, Ubuntu, so it seems to be a savfe assumption.

[1]: https://sourceware.org/glibc/wiki/ABIList

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>