fakemachine: create generic symlink for exposed images
The fakemachine CreateImage and CreateImageWithLabel functions expose an
external image to the internal machine. These functions return a
speculative path which the fakemachine must expose the images to once it
starts.

Currently images exposed to the fakemachine are exposed to a speculative
path which takes into account the virtualisation technology, which may
not be decided at the stage these functions are called. This patch allows
the functions to return a generic path which is implemented once the
machine starts running, allowing the virtualisation technology to be
decoupled.

The images and their partitions are exposed by their label under
/dev/disk/by-fakemachine-label/

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>