backend: allow choice of backend
The Machine object allows the backend name to be set using the
NewMachineWithBackend method. The backend is used when starting the
machine and allows the implementation detail of the particular
virtualisation technology to be hidden from the user.

Also for the fakemachine binary accept a cmdline parameter --backend
which allows the user to set a specific backend when calling
fakemachine manually.

By default the backend name is set to "auto" which should choose the best
supported virtualisation technology on this machine, making sure we log
the backend used. No breaking API changes are introduced.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>