backend: make auto backend fallback to uml
Currently the auto backend defaults to kvm, and if kvm is unsupported
then the auto backend is unsupported. With this patch the auto backend
chooses kvm first if supported, falling back to uml if kvm is
unsupported.

If neither backend is supported, a combined error message is shown to
the user.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>