uml: Make backend distro independant
Debian provides the UML binary as `linux.uml` and the UML modules in
`/usr/lib/uml/modules` and Slirp helper as `libslirp-helper`.

Arch Linux through AUR provides UML binaries and Slirp as `vm-linux`,
`/usr/lib/modules/*-usermodelinux` and `slirp-helper`.

The kernel documentation mentions about `linux` and the sources creates
the exact same binaries as `linux` and `vm-linux`.

This intends to make the backend less distro-specific by looking for the
first match in an hardcoded list of known files/directory.