backend: remove InitrdModules() in favour of InitModules()
With recently the former was reworked from list of modules and their
respective dependencies (all of which including relative path and
extension) to a modules that we use.

The only difference between the two is - virtio and virtio_ring. Both of
those are low level dependencies of the other virtio modules.

So just remove the extra function/array and consistently use
InitModules() across the board.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>