backend: kvm: add support for kernels in ModulePath()
Seemingly /lib/modules/$(uname -r)/vmlinuz is a thing on many modern
distributions. Perhaps this is inspired by systemd which refuses to
hibernate if the kernel isn't there.

Either way, we can use this and handle multiple distros at once.
I've confirmed this with Arch and @gasinvein has confirmed that it works
on Fedora.

As an added bonus, at least on Arch, using this approach means that we
do not need to install the initramdisk generator, which is also
responsible for copying the kernel over to /boot.

I really should have listened to @gasinvein's feedback and used this
approach in the first place.

v2: Don't add the kernel version twice
v3: Drop Debian from inline comment

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>