Check mountpoints exist on the host early
If a mountpoint doesn't exist on the host, the backend will fail to mount
it and results in an ugly error from the backend:

    $ fakemachine -v /null
    qemu-system-x86_64: -virtfs local,mount_tag=virtfs-5,path=/null,security_model=none: cannot initialize fsdev 'virtfs-5': failed to open '/null': No such file or directory
    fakemachine: error starting kvm backend: <nil>

Check if a mountpoint exists before attempting to start the backend.

While we are here also add some trivial checks against any whitespace in
the mountpoint label & machine directory which could cause issues.

Closes: #54
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>