Automatically choose backend backed on priority
When new backends are added to Fakemachine there are multiple
places in the code where the new backend must be added. Let's
reduce the ability for human errors in these cases and have
just one mapping of backend vs. the priority in which it should
be chosen in the automatic backend selection algorithm.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>