Run CI tests on user-mode-linux backend on Debian
Currently we only run the tests on the qemu backend, which
misses all of the user-mode-linux backend functionality;
specifically fakemachine setup/teardown steps. We need to be
able to reliably test all backends in CI to be able to merge
any related changes with any confidence.

For now, only testing on the Debian base containers is
supported. More container baselines may be added as follow-up
pull requests.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>