Fix TestImage unit test for user-mode-linux backend
The TestImage unit test fails for the user-mode-linux backend:

    $ go test -v -run TestImage$ ./... --backend=uml
    Running test -b /dev/vda using uml backend
    Powering off.
    --- FAIL: TestImage (6.22s)
        machine_test.go:57: Test for the virtual image device failed with 1
    FAIL
    FAIL	github.com/go-debos/fakemachine	6.222s
    ?   	github.com/go-debos/fakemachine/cmd/fakemachine	[no test files]
    ?   	github.com/go-debos/fakemachine/cpio	[no test files]
    FAIL

This is due to the test hard-coding the block device path; which
is the format used by paravirtualizated block devices, specifically a
virtio block device used by qemu/kvm. UML uses the udb driver, which
uses a different naming scheme for the block devices (e.g. /dev/ubda).

Fortunately in fakemachine we create symlinks to each disk under
/dev/disk/by-fakemachine-label/<disk_label>, which allows the user
to use a generic no matter the backend used.

Switch to using these symlinks and as such fix the test for the
user-mode-linux backend. After switching to use the symlink, the
test passes:

    $ go test -v -run TestImage$ ./... --backend=uml
    Running test -b /dev/disk/by-fakemachine-label/fakedisk-0 using uml backend
    Powering off.
    --- PASS: TestImage (6.80s)
    PASS
    ok  	github.com/go-debos/fakemachine	6.802s
    ?   	github.com/go-debos/fakemachine/cmd/fakemachine	[no test files]
    ?   	github.com/go-debos/fakemachine/cpio	[no test files]

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>