Do not automatically select the qemu backend
The qemu backend is only useful for CI testing where KVM isn't available.
Do not allow users to automatically choose this backend as it is far too
slow for regular use.

Let the CI and power-users opt-in to using this backend by manually
selecting it (either from the API or by passing --backend=qemu).

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>