Disable plymouth boot splash
If plymouth is installed in the host, the service  will be started during
the fakemachine's boot sequence. When plymouth starts, it attempts to
clear all of the the serial consoles which in turn causes the host's
shell to be cleared.

If fakemachine is ran with --show-boot this causes the boot log to be
lost and can also cause strange artifacts on the shell (e.g. missing
characters).

Add plymouth.enable=0 to the kernel arguments to stop plymouth from
starting during the machine's boot. The uml backend has the quiet parameter
which was added to achieve the same goal, so unify the kernel arguments
between the backends.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>