qemu: Explicitly depend on e1000 module
We use the e1000 module for the virtualised ethernet adaptor inside the
qemu guest. For Debian, e1000 is builtin to the kernel so we haven't
needed to include it as a dependency. When running fakemachine on Ubuntu
Jammy, fakemachine fails to setup the network as e1000 is built as a
module. Include e1000 in the list of required modules such that we support
either case: e1000 builtin or built as a module.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>