Fix systemd-networkd-wait-online race
Older versions of systemd-networkd-wait-online (e.g. <253) will happily
just exit without error if no links are available. This causes a race
condition where in some cases a fakemachine will be created with no network
interface.

Fix this by giving the network interface a proper name when it is created
and waiting for that specific network interface to go online.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>