Test image sector size
We now support a SetSectorSize method, add a test for the two
common values 512 and 4096.

The test start a machine with a given sector size and verify that
the block device correctly reports the physical and logical sizes.

Skip 4k sector size test on UML backend since it's not supported.

Signed-off-by: Julien Massot <julien.massot@collabora.com>