Create/update issue when scheduled CI job fails
It currently is quite difficult to see when a scheduled CI job (e.g. a
weekly build/test run) fails; which can cause bugs to go unnoticed for
a while. Add in a job to create an issue when the job fails, to alert any
users who watch the project's issues.

Signed-off-by: Christopher Obbard <christopher.obbard@linaro.org>